==> Modern Statistics for Modern Biology/8.Rmd <==
---
title: '8'
output: html_document
date: "2023-04-14"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## High-Throughput Count Data 

```{r}
## -----------------------------------------------------------------------------
fn = system.file("extdata", "pasilla_gene_counts.tsv",
                  package = "pasilla", mustWork = TRUE)
counts = as.matrix(read.csv(fn, sep = "\t", row.names = "gene_id"))

## -----------------------------------------------------------------------------
dim(counts)
counts[ 2000+(0:3), ]
```

```{r}
## -----------------------------------------------------------------------------
conditionNames = (sub("[[:digit:]]$", "", colnames(counts))) #$
stopifnot(length(unique(conditionNames)) == 2,
  sum(conditionNames=="untreated") == 4,
  sum(conditionNames=="treated")   == 3)
```

```{r}
## -----------------------------------------------------------------------------
szfcDemo = data.frame(
  x = c(2, 4, 6, 6,  8) * 10,
  y = c(3, 6, 2, 9, 12) * 10,
  name = LETTERS[1:5],
  check.names = FALSE)
slopes =  c(
  blue = with(szfcDemo, sum(y) / sum(x)),
  red = szfcDemo[, c("x", "y")] |> as.matrix() |>
    (DESeq2::estimateSizeFactorsForMatrix)() |> (\(x) x[2]/x[1])() |> as.vector()
)
library("ggplot2")
ggplot(szfcDemo, aes(x = x, y = y, label = name)) + geom_point() +
  coord_fixed() + xlim(c(0, 128)) + ylim(c(0, 128)) + xlab("sample 1") + ylab("sample 2") +
  geom_text(hjust= 0.5, vjust = -0.6) +
  geom_abline(slope = slopes[1], col = names(slopes)[1]) +
  geom_abline(slope = slopes[2], col = names(slopes)[2])
```

```{r}

## -----------------------------------------------------------------------------
library("tibble")
library("ggplot2")
library("DESeq2")
ggplot(tibble(
  `size factor` = estimateSizeFactorsForMatrix(counts),
  `sum` = colSums(counts)), aes(x = `size factor`, y = `sum`)) +
  geom_point()
```

```{r}
## -----------------------------------------------------------------------------
library("matrixStats")
sf = estimateSizeFactorsForMatrix(counts)
ncounts  = counts / matrix(sf,
   byrow = TRUE, ncol = ncol(counts), nrow = nrow(counts))
uncounts = ncounts[, grep("^untreated", colnames(ncounts)),
                     drop = FALSE]
ggplot(tibble(
        mean = rowMeans(uncounts),
        var  = rowVars( uncounts)),
     aes(x = log(mean), y = log(var))) +
  geom_hex() + coord_fixed() + theme(legend.position = "none") +
  geom_abline(slope = 1:2, color = c("forestgreen", "red"))
```

```{r}

## -----------------------------------------------------------------------------
annotationFile = system.file("extdata",
  "pasilla_sample_annotation.csv",
  package = "pasilla", mustWork = TRUE)
pasillaSampleAnno = readr::read_csv(annotationFile)
pasillaSampleAnno

```


```{r}
## -----------------------------------------------------------------------------
library("dplyr")
pasillaSampleAnno = mutate(pasillaSampleAnno,
condition = factor(condition, levels = c("untreated", "treated")),
type = factor(sub("-.*", "", type), levels = c("single", "paired")))
## -----------------------------------------------------------------------------
stopifnot(
  !any(is.na(pasillaSampleAnno$condition)),
  !any(is.na(pasillaSampleAnno$type)),
  sum(pasillaSampleAnno$type == "single") == 3,
  sum(pasillaSampleAnno$type == "paired") == 4)

## -----------------------------------------------------------------------------
with(pasillaSampleAnno,
       table(condition, type))
```

```{r}
## -----------------------------------------------------------------------------
mt = match(colnames(counts), sub("fb$", "", pasillaSampleAnno$file))
stopifnot(!any(is.na(mt)))

pasilla = DESeqDataSetFromMatrix(
  countData = counts,
  colData   = pasillaSampleAnno[mt, ],
  design    = ~ condition)
class(pasilla)
is(pasilla, "SummarizedExperiment")
```

```{r}

## -----------------------------------------------------------------------------
stopifnot(is(pasilla, "SummarizedExperiment"))

## -----------------------------------------------------------------------------
pasilla = DESeq(pasilla)
```

```{r}
## -----------------------------------------------------------------------------
res = results(pasilla)
res[order(res$padj), ] |> head()

## -----------------------------------------------------------------------------
ggplot(as(res, "data.frame"), aes(x = pvalue)) +
  geom_histogram(binwidth = 0.01, fill = "Royalblue", boundary = 0)
```

```{r}
## -----------------------------------------------------------------------------
thehist = hist(res$pvalue, breaks = 100, plot=FALSE)
thehist$bgl = median(thehist$counts)

## -----------------------------------------------------------------------------
plotMA(pasilla, ylim = c( -2, 2))
```

```{r}
## -----------------------------------------------------------------------------
pas_rlog = rlogTransformation(pasilla)
plotPCA(pas_rlog, intgroup=c("condition", "type")) + coord_fixed()
```

```{r}
## -----------------------------------------------------------------------------
library("pheatmap")
select = order(rowMeans(assay(pas_rlog)), decreasing = TRUE)[1:30]
pheatmap( assay(pas_rlog)[select, ],
     scale = "row",
     annotation_col = as.data.frame(
        colData(pas_rlog)[, c("condition", "type")] ))
```

```{r}
## -----------------------------------------------------------------------------
write.csv(as.data.frame(res), file = "treated_vs_untreated.csv")

## -----------------------------------------------------------------------------
rho = function(x, s)
  ifelse(abs(x) < s, x^2 / 2,  s * abs(x) - s^2 / 2)

df = tibble(
  x        = seq(-7, 7, length.out = 100),
  parabola = x ^ 2 / 2,
  Huber    = rho(x, s = 2))

ggplot(reshape2::melt(df, id.vars = "x"),
  aes(x = x, y = value, col = variable)) + geom_line()
```


 - A generalized linear model for count data  


The differential expression analysis in `DESeq2` uses a generalized linear model of the form:

 
$$K_{ij}\sim GP(\mu_{ij}, \alpha_i)$$
$$\mu_{ij}=s_jq_{ij}$$
$$\log_2(q_{ij})=\sum_{k}^{}x_{jk}\beta_{ik}$$
 
 

Let us unpack this step by step. The counts $K_{ij}$ for gene $i$, sample $j$ are modeled using a gamma-Poisson (GP) distribution with two parameters, the mean $\mu_{ij}$ and the gene specific dispersion $\alpha_i$. By default, the dispersion is different for each gene, but the same across all samples, therefore it has no index $j$. The second line in Equation states that the mean is composed of a sample-specific size factor $s_j$ and $q_{ij}$, which is proportional to the true expected concentration of fragments for gene $i$  in sample $j$. The value of $q_{ij}$
is given by the linear model in the third line via the link function, $\log_2$. The design matrix $x_{jk}$ is the same for all genes (and therefore does not depend on $i$). Its rows $j$ correspond to the samples, its columns $k$ 
correspond to the experimental factors. In the simplest case, for a pairwise comparison, the design matrix has only two columns, one of them everywhere filled with $1$ (corresponding to $\beta_0$) and the other one containing $0$ or $1$ depending on whether the sample belongs to one or the other group. The coefficients $\beta_{ik}$ give the 
$\log_2$ fold changes for gene $i$ for each column of the design matrix $X$.

```{r}
## -----------------------------------------------------------------------------
pasillaTwoFactor = pasilla
design(pasillaTwoFactor) = formula(~ type + condition)
pasillaTwoFactor = DESeq(pasillaTwoFactor)

## -----------------------------------------------------------------------------
res2 = results(pasillaTwoFactor)
head(res2, n = 3)
```

```{r}
## -----------------------------------------------------------------------------
resType = results(pasillaTwoFactor,
  contrast = c("type", "single", "paired"))
head(resType, n = 3)

```

```{r}
## -----------------------------------------------------------------------------
trsf = function(x) ifelse(is.na(x), 0, (-log10(x)) ^ (1/6))
ggplot(tibble(pOne = res$pvalue,
              pTwo = res2$pvalue),
    aes(x = trsf(pOne), y = trsf(pTwo))) +
    geom_hex(bins = 75) + coord_fixed() +
    xlab("Single factor analysis (condition)") +
    ylab("Two factor analysis (type + condition)") +
    geom_abline(col = "orange")
```

```{r}
## -----------------------------------------------------------------------------
compareRes = table(
   `simple analysis` = res$padj < 0.1,
   `two factor` = res2$padj < 0.1 )
addmargins( compareRes )

```

```{r}
## -----------------------------------------------------------------------------
stopifnot(compareRes[1, 2] > compareRes[2, 1])

## -----------------------------------------------------------------------------
# The following code is based on guidance from Mike Love
res1  = results(pasilla, cooksCutoff = FALSE)
res2  = lfcShrink(pasilla, coef = "condition_treated_vs_untreated", type="normal", res = res1)
```


```{r}


# Something like the two lines commented out below may be used to reproduce   
# how Mike Love selected the two genes: they should have similar intercepts,
# large unshrunken fold change and very different Wald statistic (i.e., have
# small / large dispersion, respectively):
#
# with(res1,
#  plot(baseMean, log2FoldChange, log = "x", ylim = c(0, 3), xlim = c(10, 1e5),
#       col = ifelse(padj < 0.1, "red", "black"), cex = log(abs(stat))))
# rownames(res1)[with(res1, identify(baseMean, log2FoldChange))]

genes = c(A = "FBgn0053926", B = "FBgn0260933")
cols  = c(FBgn0053926 = "forestgreen", FBgn0260933 = "dodgerblue3", prior = "black")

df1 = tibble(
  k         = as.vector(counts(pasilla, normalized = TRUE)[genes, ]),
  condition = rep(colData(pasilla)[["condition"]], each = length(genes)),
  gene      = rep(genes, times = ncol(pasilla)))

beta    = seq(from = -1, to = 1.5, length = 500)
kounts  = counts(pasilla)[genes,, drop = FALSE]
disps   = dispersions(pasilla[genes, ]) |> `names<-`(genes)

sf      = sizeFactors(pasilla)
cond    = as.numeric(pasilla$condition)-1 

betaPriorVar = priorInfo(res2)$betaPriorVar
priorSigma = sqrt(betaPriorVar["conditiontreated"])
prior = dnorm(beta, mean = 0, sd = priorSigma)

likelihood = function(k, alpha, intercept) {
  z = vapply(beta, function(b) {
    prod(dnbinom(k, mu = sf * 2^(intercept + b * cond), size = 1/alpha))
  }, numeric(1))
  z / (sum(z) * diff(beta[1:2]))
}

posterior = function(k, alpha, intercept) {
  z = likelihood(k, alpha, intercept) * prior
  z / (sum(z) * diff(beta[1:2]))
}

intercepts = with(mcols(pasilla[genes,]), Intercept) |> `names<-`(genes)

df2 = bind_rows(
  tibble(beta = beta, y = prior, gene = "prior", what = "pre"),
  bind_rows(
  lapply(genes, function(i) bind_rows(
    tibble(beta = beta, gene = i, what = "pre",
           y = likelihood(k = kounts[i, ], alpha = disps[i],
                          intercept = intercepts[i])),
    tibble(beta = beta, gene = i, what = "post",
           y = posterior(k = kounts[i, ], alpha = disps[i],
                         intercept = intercepts[i]))))
  )
)

is_max = function(y)
  ifelse(seq(along = y) == which.max(y), y, NA_real_)

df2 %<>% group_by(gene, what) %>% mutate(py = is_max(y))
```


```{r}
## some consistency checks:
deseqNoPrior = res1[genes, "log2FoldChange"]
deseqPrior   = res2[genes, "log2FoldChange"]
mleFromPlot  = c(beta[which.max(likelihood(kounts[1,], disps[1], intercepts[1]))],
                 beta[which.max(likelihood(kounts[2,], disps[2], intercepts[2]))])
mapFromPlot  = c(beta[which.max( posterior(kounts[1,], disps[1], intercepts[1]))],
                 beta[which.max( posterior(kounts[2,], disps[2], intercepts[2]))])
stopifnot(all(abs(deseqNoPrior - mleFromPlot) < 0.002))

## -----------------------------------------------------------------------------
library("ggbeeswarm")
ggplot(df1, aes(x = condition, y = k, col = gene)) + geom_beeswarm(cex = 5) +
      facet_grid(. ~ gene) + ylab("normalized counts") + scale_y_log10() +
      scale_color_manual(values = cols) + theme(legend.position = "none")

```

```{r}
ggplot(df2, aes(x = beta, col = gene, linetype = what)) +
  geom_line(aes(y = y)) + geom_point(aes(y = py)) +
  scale_color_manual(values = cols) + theme(legend.position = "none") +
  scale_linetype_manual(values = c(pre = "solid", post = "dotted")) +
  xlab(expression(beta)) + ylab("density")
```

```{r}
## -----------------------------------------------------------------------------
vsp = varianceStabilizingTransformation(pasilla)

## -----------------------------------------------------------------------------
j = 1
ggplot(
  tibble(
    counts = rep(assay(pasilla)[, j], 2),
    transformed = c(
      assay(vsp)[, j],
      log2(assay(pasilla)[, j])
      ),
    transformation = rep(c("VST", "log2"), each = nrow(pasilla))
  ),
  aes(x = counts, y = transformed, col = transformation)) +
  geom_line() + xlim(c(0, 600)) + ylim(c(0, 9))
```

```{r}
## -----------------------------------------------------------------------------
library("vsn")
rlp = rlogTransformation(pasilla)

msd = function(x)
  meanSdPlot(x, plot = FALSE)$gg + ylim(c(0, 1)) +
     theme(legend.position = "none")

gridExtra::grid.arrange(
  msd(log2(counts(pasilla, normalized = TRUE) + 1)) +
    ylab("sd(log2)"),
  msd(assay(vsp)) + ylab("sd(vst)"),
  msd(assay(rlp)) + ylab("sd(rlog)"),
  ncol = 3
)
```

```{r}
## -----------------------------------------------------------------------------
par(mfrow = c(4, 1), mar = c(2, 2, 1, 1))
myMA = function(h, v, theta = 0.5) {
  plotMA(pasilla, lfcThreshold = theta, altHypothesis = h,
         ylim = c(-2.5, 2.5))
  abline(h = v * theta, col = "dodgerblue", lwd = 2)
}
myMA("greaterAbs", c(-1, 1))
myMA("lessAbs",    c(-1, 1))
myMA("greater",          1)
myMA("less",         -1   )
```

```{r}
## -----------------------------------------------------------------------------
library("magrittr")
ng = 10000 # number of gene
ns = 12  # number of samples
x1 = x2 = matrix(rnorm(ns * ng), ncol = ns, nrow= ng)
seq_len(ns) %% 2
group = factor(letters[1 + seq_len(ns) %% 2])  %T>% print
group
```

```{r}
batch = factor(ifelse(seq_len(ns) <= ns/2, "B1", "B2")) %T>% print
table(group, batch)
```

```{r}
head(x1)
head(x2)
group

```

```{r}
head(genefilter::rowttests(x1, fac = group))
```
```{r}
cbind(type = "x1", genefilter::rowttests(x1, fac = group))
```

```{r}
cbind(type = "x2", genefilter::rowttests(x2, fac = group))
```

```{r}


x2[, batch=="B2"] = x2[, batch=="B2"] + 2 * rnorm(ng)
pvals = rbind(
  cbind(type = "x1", genefilter::rowttests(x1, fac = group)),
  cbind(type = "x2", genefilter::rowttests(x2, fac = group)))
ggplot(pvals, aes(x = p.value)) + 
  geom_histogram(binwidth = 0.02, boundary = 0) +
  facet_grid(type ~ .)

## -----------------------------------------------------------------------------
## library("shiny")
## shinyUI(fluidPage(
##   titlePanel("Breakdown"),
##   sidebarLayout(
##     sidebarPanel(     # select oulier shift
##       sliderInput("shift", "Outlier:", min = 0, max = 100, value = 0),
##       radioButtons("method", "Method:",
##                    c("Non-robust least squares" = "lm",
##                      "M-estimation" = "rlm"))
##     ),
##     mainPanel(       # show fit
##       plotOutput("regPlot")
##     )
##   )
## ))

## -----------------------------------------------------------------------------
## library("shiny")
## library("ggplot2")
## library("MASS")
## shinyServer(function(input, output) {
##   output$regPlot = renderPlot({
##     whpt = 15
##     mtcars_new = mtcars
##     mtcars_new$mpg[whpt] = mtcars_new$mpg[whpt] + input$shift
##     reg = switch(input$method,
##       lm = lm(mpg ~ disp, data = mtcars_new),
##       rlm = rlm(mpg ~ disp, data = mtcars_new),
##       stop("Unimplemented method:", input$method)
##     )
##     ggplot(mtcars_new, aes(x = disp, y = mpg)) + geom_point() +
##       geom_abline(intercept = reg$coefficients["(Intercept)"],
##                   slope = reg$coefficients["disp"], col = "blue")
##   })
## })

```

```{r}
head(x1)
head(x2)
group

```
```{r}
library("shiny")
shinyUI(fluidPage(
  titlePanel("Breakdown"),
  sidebarLayout(
    sidebarPanel(     # select oulier shift
      sliderInput("shift", "Outlier:", min = 0, max = 100, value = 0),
      radioButtons("method", "Method:",
                   c("Non-robust least squares" = "lm",
                     "M-estimation" = "rlm"))
    ),
    mainPanel(       # show fit
      plotOutput("regPlot")
    )
  )
))



library("shiny")
library("ggplot2")
library("MASS")
shinyServer(function(input, output) {
  output$regPlot = renderPlot({
    whpt = 15
    mtcars_new = mtcars
    mtcars_new$mpg[whpt] = mtcars_new$mpg[whpt] + input$shift
    reg = switch(input$method,
      lm = lm(mpg ~ disp, data = mtcars_new),
      rlm = rlm(mpg ~ disp, data = mtcars_new),
      stop("Unimplemented method:", input$method)
    )
    ggplot(mtcars_new, aes(x = disp, y = mpg)) + geom_point() +
      geom_abline(intercept = reg$coefficients["(Intercept)"],
                  slope = reg$coefficients["disp"], col = "blue")
  })
})
```
